# openapi/coog.yaml
openapi: 3.0.3
info:
  title: Coog API
  version: 0.1.1
  description: Control plane for the Coog media server.
servers:
  - url: http://127.0.0.1:8090
paths:
...
  /api/v1/jobs:
    get:
      summary: List acquire jobs
      responses:
        "200":
          description: Job list
    post:
      summary: Enqueue a yt-dlp job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                type:
                  type: string
                  default: ytdlp
                url:
                  type: string
                title:
                  type: string
      responses:
        "201":
          description: Job created
        "400":
          description: Unsupported type or missing URL
  /api/v1/jobs/{id}:
    get:
      summary: Job detail (progress, ready, expectedDurationMs, bufferedMs)
      parameters:
        - in: path
          name: id
          required: true
          schema:
                    type: string
      responses:
        "200":
          description: Job
        "404":
          description: Not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
security:
  - bearerAuth: []